Revert "Merge branch 'master' of https://github.com/canorusmusic/canorus"
This reverts commit 79ae52a86315cc19462a1abb72fef4e564959806, reversing
changes made to cbee22f1004bc9d2f5974727cd86cbc1deefefad.

+++ lib/rtmidi-4.0.0/RtMidi.cpp
@@ -1663,12 +1663,12 @@ static void *alsaMidiHandler( void *ptr )
           struct timespec y;
           y.tv_nsec = apiData->lastTime.tv_nsec;
           y.tv_sec = apiData->lastTime.tv_sec;
-          if ( (int)x.tv_nsec < y.tv_nsec ) {
+          if ( x.tv_nsec < y.tv_nsec ) {
               int nsec = (y.tv_nsec - (int)x.tv_nsec) / 1000000000 + 1;
               y.tv_nsec -= 1000000000 * nsec;
               y.tv_sec += nsec;
           }
-          if ( (int)x.tv_nsec - y.tv_nsec > 1000000000 ) {
+          if ( x.tv_nsec - y.tv_nsec > 1000000000 ) {
               int nsec = ((int)x.tv_nsec - y.tv_nsec) / 1000000000;
               y.tv_nsec += 1000000000 * nsec;
               y.tv_sec -= nsec;
